+++ stream_compaction/common.h
@@ -9,6 +9,7 @@
 #include <algorithm>
 #include <chrono>
 #include <stdexcept>
+#include <vector>
 
 #define FILENAME (strrchr(__FILE__, '/') ? strrchr(__FILE__, '/') + 1 : __FILE__)
 #define checkCUDAError(msg) checkCUDAErrorFn(msg, FILENAME, __LINE__)
@@ -35,6 +36,47 @@ inline int ilog2ceil(int x) {
 
 namespace StreamCompaction {
     namespace Common {
+        class devDataBuffer {
+        private:
+            int* dev_data;
+            int totalSize, size_;
+            std::vector<int> sizes;
+            std::vector<int> offsets;
+        public:
+            devDataBuffer(int n, int blockSize) :totalSize(0), size_(0) {
+                int extendedSize = BLOCKSPERGRID(n, blockSize) * blockSize;
+                while (extendedSize > 1) {
+                    if (extendedSize < blockSize) {
+                        break;
+                    }
+                    size_++;
+                    sizes.push_back(extendedSize);
+                    offsets.push_back(totalSize);
+                    totalSize += extendedSize;
+                    extendedSize = BLOCKSPERGRID(extendedSize, blockSize);
+                }
+                cudaMalloc((void**)&dev_data, sizeof(int) * totalSize);
+            }
+            ~devDataBuffer() {
+                cudaFree(dev_data);
+            }
+            int* operator[](int i) const {
+                return dev_data + offsets[i];
+            }
+            int* data() const {
+                return dev_data;
+            }
+            int size() const {
+                return size_;
+            }
+            int memCnt()const {
+                return totalSize;
+            }
+            int sizeAt(int i) const {
+                return sizes[i];
+            }
+
+        };
         __global__ void kernMapToBoolean(int n, int *bools, const int *idata);
 
         __global__ void kernScatter(int n, int *odata,
