+++ stream_compaction/radixSort.cu
@@ -45,8 +45,8 @@ namespace StreamCompaction {
             cudaMalloc((void**)&dev_falses, extended_n * sizeof(int));
             cudaMemcpy(dev_idata, idata, n * sizeof(int), cudaMemcpyHostToDevice);
             timer().startGpuTimer();
-            int maxMask = 1 << numBits;
-            for (int mask = 1; mask < maxMask; mask <<= 1)
+            unsigned maxMask = 1 << numBits;
+            for (unsigned mask = 1; mask < maxMask; mask <<= 1)
             {
                 kernMapToBoolean << <fullBlocksPerGrid, blockSize >> > (n, dev_falses, dev_idata, mask);
                 StreamCompaction::Efficient::scanInplace(extended_n, dev_falses);
@@ -62,6 +62,34 @@ namespace StreamCompaction {
             cudaFree(dev_idata);
             cudaFree(dev_odata);
             cudaFree(dev_falses);
+        }
+
+        void sortShared(int n, int* odata, const int* idata, int numBits) {
+            int extended_n = 1 << ilog2ceil(n);
+            dim3 fullBlocksPerGrid = BLOCKSPERGRID(n, blockSize);
+            int* dev_idata;
+            int* dev_odata;
+            cudaMalloc((void**)&dev_idata, n * sizeof(int));
+            cudaMalloc((void**)&dev_odata, n * sizeof(int));
+            Common::devDataBuffer buffer(extended_n, blockSize);
+            cudaMemcpy(dev_idata, idata, n * sizeof(int), cudaMemcpyHostToDevice);
+            timer().startGpuTimer();
+            unsigned int maxMask = 1 << numBits;
+            for (unsigned mask = 1; mask < maxMask; mask <<= 1)
+            {
+                kernMapToBoolean << <fullBlocksPerGrid, blockSize >> > (n, buffer.data(), dev_idata, mask);
+                StreamCompaction::Efficient::scanSharedInplace(extended_n, buffer);
+                int totalFalses = 0, tmp_back = 0;
+                cudaMemcpy(&totalFalses, buffer.data() + n - 1, sizeof(int), cudaMemcpyDeviceToHost);
+                cudaMemcpy(&tmp_back, dev_idata + n - 1, sizeof(int), cudaMemcpyDeviceToHost);
+                totalFalses += (tmp_back & mask) == 0;
+                kernScatter << <fullBlocksPerGrid, blockSize >> > (n, dev_odata, dev_idata, buffer.data(), mask, totalFalses);
+                std::swap(dev_idata, dev_odata);
+            }
+            timer().endGpuTimer();
+            cudaMemcpy(odata, dev_idata, n * sizeof(int), cudaMemcpyDeviceToHost);
+            cudaFree(dev_idata);
+            cudaFree(dev_odata);
         }
 
     }
