+++ stream_compaction/radixSort.h
@@ -7,5 +7,6 @@ namespace StreamCompaction {
         StreamCompaction::Common::PerformanceTimer& timer();
 
         void sort(int n, int* odata, const int* idata, int numBits);
+        void sortShared(int n, int* odata, const int* idata, int numBits);
     }
 }
